apphub/requirements.txt
fastapi
uvicorn
keyring
keyrings.alt
requests
GitPython
PyJWT
click
python-dotenv